Main.Recfast.Boltzmann — Function

Boltzmann(gj, gi, E, T) -> Any

Evaluate the boltzmann factor given the statistical weight as $\frac{g_j}{g_i} \exp{-E/k_b T}$

Main.Recfast.SahaBoltz — Function

SahaBoltz(gi, gc, ne, E_ion, T) -> Any

Evaluate the Saha equation given the statisical weights Calculate the Boltzmann factor for detailed balance as $\frac{n_i}{n_c} = n_e\frac{g_i}{2g_c} \left(\frac{h^2}{2 π m_e k_b T}\right)^{3/2} \exp{(E_{\rm ion}k_b T)}$ given the statistical weights, electron number density, ionization energy and temperature. All units are SI.

Main.Recfast.alphaH_func — Function

alphaH_funct(TM)

Calculate the case B recombination rate as a function of the matter temperature in m^3/s as $a_1 * 10^{-19} (T/1e4)^{a_2}/(1 + a_3 T^{a_4})$ (untested so far)

Main.Recfast.alphaHe_func — Function

Calculate the case B recomination rate for netural He in m^3/s using the formula $\frac{a_1}{sqrt{T_m/T_0} * (1 + \sqrt{T_m/ T_0})^{1-a_2}} (1+\sqrt{T_m/T_1})^{1 + a_2}$

Main.Recfast.H_z_loc — Function

H_z_loc(params, z)

We calculate the neutrino fraction as $F_{\nu} = \frac{\rho_\nu}{\rho_\gamma} = N_{eff} * \frac7 8 * \left(\frac{4}{11}\right)^{4/3},$

the redshfit at matter radiation equality as $z_{eq} = \frac{3 * (H_0 c)^2 * \Omega_m }{8 \pi G a (1 + F_nu) * T_0^4 } - 1,$

where $a$ is the radiation constant and $T_0$ is the present temperature of the CMB. Then,

$H(z) = H_0 \sqrt{\Omega_\Lambda + (1 + z)^2 \left[\Omega_K + (1+z) \Omega_M \left(1 + \frac{1 +z}{1+ z_{eq}}\right)\right]}$

Note, $H(z)$ is given in $\rm s^{-1}$

Main.Recfast.NH — Function

NH(params, z)`

We calculate the hydrogen number density in $\rm m^{-3}$ as:

$3 H_0^2 \frac{\Omega_B}{8 \pi G m_H \mu_H} (1+z)^3,$ where $\mu_H$ is $1/(1-Y_p)$.

Main.Recfast.TCMB — Function

TCMB(params, z)`

Return $T_0 * (1 + z)$

Main.Recfast.calc_Orel — Function

calc_Orel(TCMB0, Nnu, h100)`

Calculate the relativistic contribution as

$\frac{a T_0^4/c^2}{3H_0^2 / (8 \pi G)} (1 + \frac78 \left(\frac{4}{11}\right)^{4/3}N_{eff} ).$

This is used to calculate the dark energy when not specified.

Main.Recfast.Variables_Fund_Consts — Type

Struct to control the rescaling. Fields:

mode

Default: 0

RF_aScale

Default: 1.0

RF_mScale

Default: 1.0

fT

Default: 1.0

fl

Default: 1.0

fA2g

Default: 1.0

fC

Default: 1.0

fAp

Default: 1.0

fBp

Default: 1.0

fm

Default: 1.0

Main.Recfast.Params — Type

Struct hodling all the cosmology parameters. Fields:

Yp

Default: 0.24

T0

Default: 2.725

Omega_M

Default: 0.26

Omega_B

Default: 0.044

Omega_K

Default: 0.0

h100

Default: 0.71

n_eff

Default: 3.04

F

Default: 1.14

f_ann

Default: 1.0e-24

correction

Default: true

b0

Default: 3.0

nb

Default: -2.9

rescaling

Default: VariablesFundConsts()

fHe

Default: Yp / (4.0 * RFfacmHemH * (1 - Yp))

H0

Default: (h100 * 100 * 100000.0) / RF_Mpc

Omega_L

Default: ((1.0 - OmegaK) - OmegaM) - calcOrel(T0, neff, h100)

Main.Recfast.Set_VFC_params! — Function

Set_VFC_params!(r::Main.Recfast.Variables_Fund_Consts; aS, mS, mode)

Function to set the parameters fields. Mode: 0 - no rescaling

1 - Rescale boltzmann factor exponentials (temperatures)

2 - Rescale Thomson cross section

3 - Rescale 2s1s 2 photon rate

4 - Rescale alpha and beta coefficients

5 - Rescale Lyman alpha rates

6 - Rescale everything

Set_VFC_params!(p::Main.Recfast.Params; aS, mS, mode)

Main.Recfast.Evaluate_recombination — Function

Evaluate_recombination(p::Main.Recfast.Params; zstart, zend) -> Any

Evaluate the recombination history. Note that we assume everything is completely ionized at zstart. If this is not the case, we will get wrong results.