Main.Recfast.Boltzmann — Function

Boltzmann(gj, gi, E, T)

Evaluate the boltzmann factor given the statistical weight as $\frac{g_j}{g_i} \exp{-E/k_b T}$

Main.Recfast.SahaBoltz — Function

SahaBoltz(gi, gc, ne, E_ion, T)

Evaluate the Saha equation given the statisical weights Calculate the Boltzmann factor for detailed balance as $\frac{n_i}{n_c} = n_e\frac{g_i}{2g_c} \left(\frac{h^2}{2 π m_e k_b T}\right)^{3/2} \exp{(E_{\rm ion}k_b T)}$ given the statistical weights, electron number density, ionization energy and temperature. All units are SI.

Main.Recfast.alphaH_func — Function

alphaH_funct(TM)

Calculate the case B recombination rate as a function of the matter temperature in m^3/s as $a_1 * 10^{-19} (T/1e4)^{a_2}/(1 + a_3 T^{a_4})$ (untested so far)

Main.Recfast.H_z_loc — Function

Hzloc(params, z)

We calculate the neutrino fraction as $F_{\nu} = \frac{\rho_\nu}{\rho_\gamma} = N_{eff} * \frac7 8 * \left(\frac{4}{11}\right)^{4/3},$ the redshfit at matter radiation equality as $z_{eq} = \frac{3 * (H_0 c)^2 * \Omega_m }{8 \pi G a (1 + F_nu) * T_0^4 } - 1,$ where $a$ is the radiation constant and $T_0$ is the present temperature of the CMB. Then, $H(z) = H_0 \sqrt{\Omega_\Lambda + (1 + z)^2 \left[\Omega_K + (1+z) \Omega_M \left(1 + \frac{1 +z}{1+ z_{eq}}\right)\right]}$ Note, $H(z)$ is given in $\rm s^{-1}$